# src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This API provides all the functions needed to manage \nthe JABC HR\
    \ system.\n"
  version: "1.0.0"
  title: "JABC"
  contact:
    email: "jabraham9719@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/JABC/1.0.0"
...
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /document/{id}:
    get:
      tags:
      - "Document"
      summary: "gets an specific Document"
      description: "Will return the Document that matches with the provided [id] from\
        \ \nthe Employee with [id]\n"
      operationId: "getDocument"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched Document"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IDocument"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
    put:
      tags:
      - "Document"
      summary: "updates the Document"
      description: "Will update an Document with the provided data in body \nif the\
        \ Document matches the [id]\n"
      operationId: "updateDocument"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched Document"
        required: true
        type: "integer"
        format: "uint64"
      - in: "body"
        name: "document"
        description: "Document data"
        required: true
        schema:
          $ref: "#/definitions/IDocument"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
    delete:
      tags:
      - "Document"
      summary: "deletes Document"
      description: "Will delete the Document that matches with the provided [id] from\
        \ \nthe Employee with [id]\n"
      operationId: "deleteDocument"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched Document"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
  /document/type:
    get:
      tags:
      - "Document"
      summary: "get all the DocumentTypes"
      description: "This returns all the DocumentTypes of the system.\n"
      operationId: "getDocumentTypes"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/IDocumentType"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
    post:
      tags:
      - "Document"
      summary: "creates a new DocumentType"
      description: "Will create a new DocumentType with the provided data in body"
      operationId: "createDocumentType"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - in: "body"
        name: "documentType"
        description: "DocumentType data"
        required: true
        schema:
          $ref: "#/definitions/IDocumentType"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
  /document/type/{id}:
    get:
      tags:
      - "Document"
      summary: "gets an specific DocumentType"
      description: "Will return the DocumentType that matches with the provided [id]"
      operationId: "getDocumentType"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched DocumentType"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IDocumentType"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
    put:
      tags:
      - "Document"
      summary: "updates the DocumentType"
      description: "Will update an DocumentType with the provided data in body if\
        \ the DocumentType matches the [id]"
      operationId: "updateDocumentType"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the DocumentType to be updated"
        required: true
        type: "integer"
        format: "uint64"
      - in: "body"
        name: "documentType"
        description: "DocumentType data"
        required: true
        schema:
          $ref: "#/definitions/IDocumentType"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
    delete:
      tags:
      - "Document"
      summary: "deletes DocumentType"
      description: "Will delete an DocumentType if the DocumentType matches the [id]"
      operationId: "deleteDocumentType"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the DocumentType to be deleted"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Document"
...
securityDefinitions:
  AuthToken:
    type: "apiKey"
    name: "X-Auth-Token"
    in: "header"
definitions:
  IApiResponse:
    type: "object"
    required:
    - "message"
    - "responseCode"
    - "type"
    properties:
      responseCode:
        type: "integer"
        format: "int32"
      type:
        type: "string"
        enum:
        - "ERROR"
        - "SUCCESS"
      message:
        type: "string"
        example: "Unauthorized access to the API"
      debugMessage:
        type: "string"
        example: "This is a debug message"
    example:
      debugMessage: "This is a debug message"
      type: "ERROR"
      message: "Unauthorized access to the API"
      responseCode: 0
...
  IDocument:
    type: "object"
    required:
    - "createdDate"
    - "dueDate"
    - "fkDocumentType"
    - "fkEmployee"
    - "id"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the Role"
      fkDocumentType:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The foreign key  of the DocumentType of this Document"
      fkEmployee:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The foreign key  of the Employee that needs to upload this Document"
      createdDate:
        type: "integer"
        format: "uint64"
        description: "The unix timestamp of the created date of this Document"
      dueDate:
        type: "integer"
        format: "uint64"
        description: "The unix timestamp of the due date of this Document"
      expiryDate:
        type: "integer"
        format: "uint64"
        description: "The unix timestamp of the expiry date of this Document"
      path:
        type: "string"
      type:
        $ref: "#/definitions/IDocumentType"
    description: "A Document required for the onbarding proccess of a new Employee\n"
    example:
      expiryDate: 1
      path: "path"
      createdDate: 0
      fkDocumentType: 1
      dueDate: 6
      fkEmployee: 1
      id: 1
      type:
        path: "path"
        name: "name"
        description: "description"
        id: 1
  IDocumentType:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the DocumentType"
      name:
        type: "string"
      description:
        type: "string"
      path:
        type: "string"
    description: "A DocumentType contains info of the template support document for\
      \ the onboarding process\n"
    example:
      path: "path"
      name: "name"
      description: "description"
      id: 1
...